Fix error fields KeyError in message object

+++ bot_discord/bot/manage/channel_data.py
@@ -1,4 +1,4 @@
-from bot.colors import yellow
+from bot.colors import red, yellow
 
 async def flush(bot, selected_channel):
     for server in bot.servers:
@@ -8,12 +8,13 @@ async def flush(bot, selected_channel):
                 continue
             async for m in bot.logs_from(channel):
                 try:
-                    if m.embeds:
+                    if m.embeds and 'fields' in m.embeds[0].keys():
                         title = m.embeds[0]['fields'][0]['name']
                         if 'New challenge solved by' not in title:
                             await bot.delete_message(m)
                     else:
                         await bot.delete_message(m)
                 except Exception as exception:
+                    red(exception)
                     return False
     return True
